Refactor: use same component for UniqueValues and IntervalLegend in Rasters (#1235)
* refactor: removed use of `RasterUniqueValuesLegend` and use `RasterIntervalsLegend` for raster

* fix: restoring of map style for unique values

* remaned `RasterIntervalsLegend.svelte` to `RasterAdvancedLegend.svelte`
