+++ sites/geohub/src/components/controls/RasterLegendContainer.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
   import { slide } from 'svelte/transition'
   import RasterContinuousLegend from '$components/controls/RasterContinuousLegend.svelte'
-  import RasterIntervalsLegend from '$components/controls/RasterIntervalsLegend.svelte'
-  import RasterUniqueValuesLegend from '$components/controls/RasterUniqueValuesLegend.svelte'
-  import { DynamicLayerLegendTypes, COLOR_CLASS_COUNT_MAXIMUM, ClassificationMethodTypes } from '$lib/constants'
-  import type { Layer, RasterTileMetadata, IntervalLegendColorMapRow } from '$lib/types'
+  import RasterAdvancedLegend from '$components/controls/RasterAdvancedLegend.svelte'
+  import { DynamicLayerLegendTypes, ClassificationMethodTypes } from '$lib/constants'
+  import type { Layer, IntervalLegendColorMapRow } from '$lib/types'
   import LegendTypeSwitcher from './LegendTypeSwitcher.svelte'
 
   export let layer: Layer
@@ -16,28 +15,14 @@
 
   let rasterLegendType: DynamicLayerLegendTypes
 
-  let { info }: Layer = layer
-
   let colorPickerVisibleIndex: number
-  let layerHasUniqueValues = false
   let showTooltip = false
 
   $: legendType, handleLegendToggleClick()
   const handleLegendToggleClick = () => {
     colorPickerVisibleIndex = -1
-    let bandName
-
-    try {
-      bandName = Object.keys((info as RasterTileMetadata).stats)
-    } catch (e) {
-      console.log(e)
-    }
-    layerHasUniqueValues =
-      Number((info as RasterTileMetadata).stats[bandName]['unique']) <= COLOR_CLASS_COUNT_MAXIMUM &&
-      !(info as RasterTileMetadata).dtype.startsWith('float')
-
     if (legendType === 'advanced') {
-      rasterLegendType = layerHasUniqueValues ? DynamicLayerLegendTypes.UNIQUE : DynamicLayerLegendTypes.INTERVALS
+      rasterLegendType = DynamicLayerLegendTypes.INTERVALS
     } else {
       rasterLegendType = DynamicLayerLegendTypes.CONTINUOUS
     }
@@ -55,19 +40,12 @@
   </div>
 {:else if rasterLegendType === DynamicLayerLegendTypes.INTERVALS}
   <div transition:slide>
-    <RasterIntervalsLegend
+    <RasterAdvancedLegend
       bind:layerConfig={layer}
       bind:colorMapName
       bind:classificationMethod
       bind:numberOfClasses
       bind:colorMapRows
       bind:generateCmap={showTooltip} />
-  </div>
-{:else if rasterLegendType === DynamicLayerLegendTypes.UNIQUE}
-  <div transition:slide>
-    <RasterUniqueValuesLegend
-      bind:layerConfig={layer}
-      bind:colorPickerVisibleIndex
-      bind:colorMapName />
   </div>
 {/if}
